Remove test data exclusion from .gitignore.
The prototype to minespex used data from NREL, but this data is quite
large (5-10 GB). During development, the tests/data folder was excluded
from the repo to prevent the repo from becoming too large.

This had the inevitable effect of excluding the simplified data created
to test the Spectra class and the Scienta read function. This purpose
of this commit is simple--include the smaller test data into the repo.